Deep Learning/Human Detection inside a Swimming Pool/requirements.txt
opencv-python
numpy
Flask
PyYAML